% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_products.R
\name{count_products}
\alias{count_products}
\title{Count number of products}
\value{
integer indicating the number of products
}
\description{
Count number of products
}
\details{
This function will gather all product listed in the shopify website and count how many are they
}
