% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_version.R
\name{get_api_version}
\alias{get_api_version}
\title{Get Latest Stable API Version}
\value{
character indicating the latest stable API version
}
\description{
Retrieve the latest stable api version
}
\details{
Shopify releases a new API version every 3 months at the beginning of the quarter. This function follows this rule to automatically infer the latest
}
\references{
\url{https://help.shopify.com/en/api/versioning#the-api-version-release-schedule}
}
