% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Get request}
\arguments{
\item{request_url}{URL to be sent}

\item{parameters}{any specific parameters to be passed unto the 'query' parameter}
}
\description{
Get request
}
\details{
Generic method that sends a GET type of request to the rest API
}
