% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShopifyConnection.R
\docType{class}
\name{ShopifyConnection}
\alias{ShopifyConnection}
\title{This is a ShopifyConnection Class}
\description{
This type of object will open a connection with a determined shopify app
}
\details{
Keep it secret, keep it safe
}
\section{Public fields}{


* `test`: this is a test yet.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShopifyConnection$new()}}
\item \href{#method-clone}{\code{ShopifyConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShopifyConnection$new(shop_url, api_key, api_password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shop_url}}{shop URL (e.g. 'https://blue-whale-coffee.myshopify.com')}

\item{\code{api_key}}{API key}

\item{\code{api_password}}{API password}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
if successful, will return a new \code{ShopifyConnection} object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShopifyConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
